requirements.txt
amqp==5.0.3
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
prompt-toolkit==3.0.14
python-rake==1.5.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wikipedia==1.4.0
psycopg2-binary
pytest